+++ src/sdk/PnP.Core.Test/SharePoint/ListTests.cs
@@ -2364,6 +2364,43 @@ namespace PnP.Core.Test.SharePoint
         }
 
 
+        [TestMethod]
+        public async Task ItemOpenListSettingsTest()
+        {
+            //TestCommon.Instance.Mocking = false;
+            using (var context = await TestCommon.Instance.GetContextAsync(TestCommon.TestSite))
+            {
+                // Add a new library
+                string listTitle = TestCommon.GetPnPSdkTestAssetName("ItemOpenListSettingsTest");
+                IList myList = null;
+                try
+                {
+                    myList = context.Web.Lists.GetByTitle(listTitle);
+
+                    if (TestCommon.Instance.Mocking && myList != null)
+                    {
+                        Assert.Inconclusive("Test data set should be setup to not have the list available.");
+                    }
+
+                    if (myList == null)
+                    {
+                        myList = await context.Web.Lists.AddAsync(listTitle, ListTemplateType.DocumentLibrary);
+                    }
+
+                    myList.DefaultItemOpenInBrowser = false;
+                    await myList.UpdateAsync();
+
+                    myList = context.Web.Lists.GetByTitle(listTitle, p => p.DefaultItemOpenInBrowser);
+
+                    Assert.IsTrue(myList.DefaultItemOpenInBrowser == false);
+                }
+                finally
+                {
+                    myList.Delete();
+                }
+            }
+        }
+
         #endregion
     }
 }