+++ src/sdk/PnP.Core/Model/SharePoint/Core/Internal/List.cs
@@ -150,7 +150,9 @@ namespace PnP.Core.Model.SharePoint
         public string DefaultNewFormUrl { get => GetValue<string>(); set => SetValue(value); }
 
         public string DefaultViewUrl { get => GetValue<string>(); set => SetValue(value); }
-        
+
+        public bool DefaultItemOpenInBrowser { get => GetValue<bool>(); set => SetValue(value); }
+
         public ListReadingDirection Direction { get => GetValue<ListReadingDirection>(); set => SetValue(value); }
 
         public string ImageUrl { get => GetValue<string>(); set => SetValue(value); }